refactor: ModalTitle 컴포넌트 Modal.tsx에서 분리
